requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pyaml==20.4.0
PyYAML==5.4.1
Werkzeug==1.0.1
